fix(archive.php): fixed the bug that the archive page didn't show correctly the second post
Moved old archive.php to old-archive.php in the previous commit and now recreated the archve page
from scratch. The problem was created by the custom field and how it displayed the date therefore
that was removed in order to prevent any further issues. Will add that feature possibly in the
future, at the moment the archive should be added in the event title.

BREAKING CHANGE: Fixed the issue where it didn't display the second post